requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
